DDF-3815 Moved EmbeddedListMetacardsHandlerTest and fixed formatting and import issues.
EmbeddedListMetacardsHandlerTest created by a rebased commit needed to be moved to the new package for EmbeddedListMetacardsHandler (*.transformer.impl instead of *.transformations). In addition, some imports broken by moved classes were fixed and the formatter fixed some formatting issues.
