+++ catalog/ui/search-ui-app/pom.xml
@@ -97,6 +97,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>ddf.ui</groupId>
+            <artifactId>catalog-ui-search-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>ddf.ui.search</groupId>
             <artifactId>standard</artifactId>
             <version>${project.version}</version>
